Added rest of Cornell Box triangles. Changed Color to use floats instead of ints for rgb.

+++ src/Color.h
@@ -2,9 +2,9 @@
 #define _COLOR_H_
 
 typedef struct Color {
-    unsigned int red;
-    unsigned int green;
-    unsigned int blue;
+    float red;
+    float green;
+    float blue;
 } Color;
 
 #endif
